requirements.txt
Flask==3.1.2
gunicorn==25.1.0
python-dotenv==1.2.1

